move umask code into its own file
